requirements.txt
annotated-types     0.7.0
anyio               4.9.0
beautifulsoup4      4.12.3
blis                1.3.0
catalogue           2.0.10
certifi             2025.4.26
charset-normalizer  3.4.2
click               8.2.1
cloudpathlib        0.21.1
confection          0.1.5
cymem               2.0.11
DAWG-Python         0.7.2
DAWG2-Python        0.9.0
docopt              0.6.2
h11                 0.16.0
httpcore            1.0.9
httpx               0.28.1
idna                3.10
Jinja2              3.1.6
joblib              1.5.1
langcodes           3.5.0
language_data       1.3.0
marisa-trie         1.2.1
markdown-it-py      3.0.0
MarkupSafe          3.0.2
mdurl               0.1.2
murmurhash          1.0.13
nltk                3.9.1
numpy               2.3.0
packaging           25.0
pandas              2.3.0
pip                 24.0
preshed             3.0.10
pydantic            2.11.7
pydantic_core       2.33.2
Pygments            2.19.1
pymorphy2           0.9.1
pymorphy2-dicts-ru  2.4.417127.4579844
pymorphy3           2.0.4
pymorphy3-dicts-ru  2.4.417150.4580142
python-dateutil     2.9.0.post0
python-telegram-bot 22.1
pytz                2025.2
regex               2024.11.6
requests            2.31.0
rich                14.0.0
ru_core_news_sm     3.8.0
scikit-learn        1.7.0
scipy               1.15.3
setuptools          80.9.0
shellingham         1.5.4
six                 1.17.0
smart-open          7.1.0
sniffio             1.3.1
soupsieve           2.7
spacy               3.8.7
spacy-legacy        3.0.12
spacy-loggers       1.0.5
srsly               2.5.1
thinc               8.3.6
threadpoolctl       3.6.0
tqdm                4.67.1
typer               0.16.0
typing_extensions   4.14.0
typing-inspection   0.4.1
tzdata              2025.2
urllib3             2.4.0
wasabi              1.1.3
weasel              0.4.1
wrapt               1.17.2